+++ core/src/main/scala/zio/config/ConfigSourceModule.scala
@@ -11,8 +11,6 @@ import zio.config.PropertyTree.{ unflatten, Leaf, Record, Sequence }
 import java.io.FileInputStream
 import java.io.File
 
-import These._
-
 trait ConfigSourceModule extends KeyValueModule {
 
   case class ConfigSourceName(name: String)
@@ -497,6 +495,27 @@ trait ConfigSourceStringModule extends ConfigSourceModule {
       case class Value(value: String)
 
       type KeyValue = These[Key, Value]
+
+      import These._
+
+      sealed trait These[+A, +B] { self =>
+        def fold[C](
+          f: (A, B) => C,
+          g: A => C,
+          h: B => C
+        ): C = self match {
+          case This(left)        => g(left)
+          case That(right)       => h(right)
+          case Both(left, right) => f(left, right)
+        }
+      }
+
+      object These {
+        final case class Both[A, B](left: A, right: B) extends These[A, B]
+        final case class This[A](left: A)              extends These[A, Nothing]
+        final case class That[B](right: B)             extends These[Nothing, B]
+      }
+
 
       object KeyValue {
         def mk(s: String): Option[KeyValue] =
